fix fallback deployment service conf
